+++ src/main/java/cart/controller/AdminController.java
@@ -1,15 +1,15 @@
 package cart.controller;
 
 import cart.service.ProductService;
-import cart.service.dto.ProductSearchResponse;
+import cart.service.dto.product.ProductSearchResponse;
 import java.util.List;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-@Controller
 @RequestMapping("/admin")
+@Controller
 public class AdminController {
 
     private final ProductService productService;
