+++ src/main/java/cart/controller/ProductController.java
@@ -1,8 +1,8 @@
 package cart.controller;
 
 import cart.service.ProductService;
-import cart.service.dto.ProductModifyRequest;
-import cart.service.dto.ProductRegisterRequest;
+import cart.service.dto.product.ProductModifyRequest;
+import cart.service.dto.product.ProductRegisterRequest;
 import java.net.URI;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -14,8 +14,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@RestController
 @RequestMapping("/products")
+@RestController
 public class ProductController {
 
     private final ProductService productService;
