+++ src/main/java/cart/service/ProductService.java
@@ -4,9 +4,9 @@ import cart.dao.product.ProductDao;
 import cart.dao.product.ProductEntity;
 import cart.domain.product.Product;
 import cart.global.exception.ProductNotFoundException;
-import cart.service.dto.ProductModifyRequest;
-import cart.service.dto.ProductRegisterRequest;
-import cart.service.dto.ProductSearchResponse;
+import cart.service.dto.product.ProductModifyRequest;
+import cart.service.dto.product.ProductRegisterRequest;
+import cart.service.dto.product.ProductSearchResponse;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
